{{cookiecutter.project_slug}}/requirements/base.txt
argon2-cffi==19.2.0

# Django
# ------------------------------------------------------------------------------
django==3.0.5
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-filter==2.2.0

# Postgres
psycopg2==2.8.5 --no-binary psycopg2

# Docs
coreapi==2.3.3
drf-yasg==1.17.0